traverse graph at most once when marking partial flows
Summary:
### Overall problem

We had a big performance problem where the time to compute partial flows appears to scale linearly with the number of issues, causing experiments to increase the number to blow up in postprocessing. The problem is that we were searching through the same set of frames a large # of times, even if a significant subgraph of our trace graph is shared across issues, and we'll always find the same answer.

### High-level fix

To fix this, change how we mark partial flows to go from a BFS to a DFS, and memoize the DFS result. For each frame, the DFS now evaluates to whether we added a breadcrumb in any transitively reachable frame, and use this information to add breadcrumbs at the issue level as appropriate.

This guarantees we traverse each frame at most once.

### Implementation Details

I don't do a traditional recursive DFS here + collect results, because Python's recursion limit is very low (1000) by default, and Python stack frames are *really* expensive. Instead, I chose an approach that's morally similar to a topological sort with an explicit stack. The idea is that we visit every node in the graph twice, with the goal of answering the question of "did we add a partial flow feature for any transitively reachable node" for each node in the graph.

In the first pass, we simply queue a first pass for all the node's direct children, followed by a second pass for the node itself, and continue. We take care to make sure the first pass only happens once.

The second pass assumes that we know whether any reachable frame's results already, and updates the data structure with the final answer for the current node.

In practice, this is pretty fast, and took ~4 seconds to mark all partial flows for a production usecase.

Reviewed By: fahndrich

Differential Revision: D70910942

fbshipit-source-id: a1f90eea8ee589f882b1143ddfb3c2992fa8398c