Update the parser to use the filename at the root of the model
Summary:
The pysa taint parser currently relies on each taint frame (i.e pre- and postconditions) to have a filename. This is unnecessary, since within a given model, all preconditions and postconditions will have the same filename. We could save in memory usage by removing the filename from the taint representation.
This diff removes the logic that relies on the filename in each frame, and instead uses the filename at the root of the model.

Reviewed By: tianhan0

Differential Revision: D71312596

fbshipit-source-id: 47beba5d9d71577e10983c784140da63ec14b77d