Fix errors in parser when parsing object and attribute models
Summary:
# Problem
In D71134905, we removed the module (which is used to find the filename) from the `CallInfo` part of the taint representation, assuming the module/filename was already specified at the root of the model.
Unfortunately, it looks like this is not always the case. SAPP started silently erroring when parsing our taint output, for instance:

```
2025-03-18 08:07:24,265 [ERROR] Unable to parse trace for the following:
{'callable': 'Obj{xxx[redacted]}',
 'model_generators': ['[redacted]'],
 'sources': [{'port': 'result',
              'taint': [{'declaration': None,
                         'kinds': [{'kind': 'UserControlled'}]}]}]}
```

There are two problems:
* (1) Models for attributes and global variables don't have a filename. This is expected, since those models can't have sources or sinks with a `CallInfo.Origin` or `CallInfo.Declaration` (only `CallInfo.Declaration`). The problem is that the parser doesn't know that.
* (2) Models for functions or methods sometimes don't have a filename.

# This diff
This diff solves problem (1).

Reviewed By: sinancepel

Differential Revision: D71492011

fbshipit-source-id: 111ef2121a570214f765695d7430cc72462fc01f