Handle all anonymous class prefixes to fix duplicate issues with different handles
Summary:
Duplicate issues were reported in perprator. See https://fb.workplace.com/groups/perpetrato/permalink/1534667383871645/ + examples in T219860184.
One of the root cause is because the sink `callee` which is a part of our handles are **synthetic** classes/lambdas. Since these are generated, the names are random and we strip the numbering that are a part of the name for consistency. But currently we only check for the simple java anonymous class naming convention of `$n`. The duplicate issues however are due to the synthetic classes generated in kotlin which have a different naming convention. See: https://r8.googlesource.com/r8/+/refs/tags/3.1.34/src/main/java/com/android/tools/r8/synthesis/SyntheticNaming.java#140

This diff fixes the issue by handling 2 other synthetic names for lambdas: `"$$ExternalSyntheticLambda"` and `"$$Lambda$"` in addition to the `$`. Preserves the `$` sign for tracking different nested levels so is compatible with the current implementation for `$`.

Reviewed By: GerbenJavado

Differential Revision: D72271863

fbshipit-source-id: d05e8abcb96ce9105fbbbb21b038e7a123934c26