Update inline message when a subtrace is present
Summary:
MT doesn't track the leaf name in the `CallInfo` and instead relies on the nested `Origin` to retrieve it. Since `ExtraTraces` also uses `CallInfo` but it does not track the `Origin`, we do not have a `leaf_name` available when the call_info in Origin and hence we do not add subtraces for these. The trace tree UI can be confusing because of this as the node indicates there is a subtract but one is not visible.

This diff just updates the message on the UI to explicitly state when the subtrace present to hopefully avoid some confusion.

Reviewed By: yuhshin-oss

Differential Revision: D72356476

fbshipit-source-id: 4f0a8aa076d7a833707d05e7a32c2d39546df536