Use a dataclass instead of a TypedDict for Summary
Summary:
TypedDict seems to lead to a lot of holes in the type system. For instance, type checking did not detect the open source breakage in D73654054.
Let's use `dataclass` instead of `TypedDict`, this should improve typing around the codebase.
The downside is that we use `Summary` has a mutable type, and now we will need `none_throws` in many places.

Reviewed By: yuhshin-oss

Differential Revision: D73655806

fbshipit-source-id: 014534924920298ade5f94d1493766c62774bed4