Fix partial issue search
Summary:
Looks like the partial issue trace traversal isn't correct. It uses a visited set for frames that is shared among all different full issue codes. This is obviously not correct as we need to know the partial flow spec context when we do the trace traversal.

This diff changes the visited state so that we keep one visited set per combination of `(partial_issue_code, prefix, transform)`

Reviewed By: alexblanck

Differential Revision: D74407002

fbshipit-source-id: 8f8bfe59d965dcb384d9337af8b11a71f9f16ee2