[sapp] fix sqlalchemy cache_ok spew
`sapp analyze` was spewing SQLAlchemy warnings; looks like they added
a caching attribute that must explicitly be set to True or False:

```
SAWarning: TypeDecorator IssueBIGDBIDType() will not produce
a cache key because the ``cache_ok`` attribute is not set to
True.  This can have significant performance implications
including some performance degradations in comparison to
prior SQLAlchemy versions.  Set this attribute to True if
this type object's state is safe to use in a cache key, or
False to disable this warning. (Background on this error at:
https://sqlalche.me/e/14/cprf)
```

The resolution is to add the attribute; seems safe so setting
to True.

Tested by running `sapp analyze` on results from the Zulip codebase;
seems to eliminate the spew.