Fix missing frames in issues that loop on a callable
Summary:
If two consecutive trace frames are in the same callable (but different port), the UI will be broken and shows "missing trace frame".
This is because we have an unnecessary check for `caller_id != callee_id`.
We can safely delete this since the trace frame walker already keeps trace of visited trace frames, so it should always converge.

Reviewed By: anwesht

Differential Revision: D77733224

fbshipit-source-id: cd26fd937fafea2574386f7dc00653051619d322