Fix broken traces for transforms in the open source UI
Summary:
We don't currently show transforms correctly in the open source UI, since we expect the set of leaf kinds to stay the same when traversing frames.
Instead, we need to add the logic to match the caller and callee leaf kind based on `@`.
This is what this diff does.

Reviewed By: anwesht

Differential Revision: D77738966

fbshipit-source-id: cc4eda766b33dad6cf04320008c965092f0d3c3d