Choose parser based on metadata, not command-line flags
Summary:
As we start to migrate SAPP tools over to Rust-based parsers, I wanted to clean up the way we actually select which parser to use.

When running analysis, we already parse the tool information out of the metadata and use to decide which `tool` to write to the new row in the `runs` table.

However, when selecting the parser to use, we currently use the `tool` value for some tools but command-line flags like `--pysa` and `--fontainebleau` for other tools. This means there's a potential for disagreement between flags and metadata.

This diff standardizes things so we always use the metadata's tool value and no longer require command-line flags to detect a tool.

Reviewed By: scottblaha

Differential Revision: D79309089

fbshipit-source-id: fc09e08d27ead8c4dffd8c45ebd2cdf07ef8a402