Guard against writing provisional DBID values
Summary:
Currently, `DBID`s can assigned multiple times. This risks a situation where a DBID is written to the database but then the value is reassigned later - creating broken references in the database.

To prevent this, this diff adds a "frozen" state to `DBID` to indicate that the final value has been set. Reassigning the value is prohibited when the DBID is frozen and retrieving the inner value in most cases will require that the value is frozen.

Reviewed By: scottblaha

Differential Revision: D79676263

fbshipit-source-id: def23babc341fe02ed41772ba6c4f031910c5025