Use sharded ids for RunOrigin.id
Summary:
Although we have migrated existing RunOrigin records to use sharded ids, and set the autoincrement value there (see D84556427), it would be better if we use the regular id allocation mechanism SAPP uses for all its other tables.

This must be guarded behind a JK because switching it on needs to happen somewhat atomically: we will see failures if some code is writing using autoincrement and some is trying to allocate ids.

Reviewed By: fahndrich

Differential Revision: D84552319

fbshipit-source-id: b71f37e47326d4ebed833e6c4496daa9d852d316