Fix issue deduplication bug for meta runs
Summary:
# Context
The command line parameter `-meta-run-skip-duplicate-issues` triggers an additional step that skips storing issues if they were found in another run in the same metarun.

# Problem
This is implemented by checking if the issue instance hash already exists in the database.
However, there is a bug: Imagine a run A starts uploading, updates the `metarun_issue_instance_index` index with its issues, then crashes. Then, a run B with similar issues would see those issues in `metarun_issue_instance_index` and thus skip storing the same issue. However, since the run A crashed, it will be purged by our database purging logic. In the end, those issues would not be stored in the database.

# Solution
We should only skip issues if we see that the run was marked as finished, which means uploading succeeded.

Reviewed By: tianhan0

Differential Revision: D87545289

fbshipit-source-id: 595744830454112d3950a7384c1b6861246cff62