Support "__SYNTHETIC" as a path prefix for synthetic frames
Summary: After discussing with anwesht, we decide to make the path of the synthetic methods in Mariana Trench the form of `"__SYNTHETIC:/path/to/file"` for easier debugging processes (also see D87105305 and D87108977 for more context). Therefore, this diff updates the Mariana Trench SAPP parser to handle this new file name format. The parse would drop the actual path and only keep the `"__SYNTHETIC"` prefix, such that SAPP would correctly treat it as zero-length (see D86442643 for details on zero-length frames).

Reviewed By: anwesht

Differential Revision: D87287398

fbshipit-source-id: 6831a65d132e53020b20be23525bcb116462c225