Fix exploitability root parsing in mariana trench python parser
Summary:
While porting this parser to rust, we found a bug in the way exploitability roots were being parsed:

The code was incorrectly assuming that the roots are aways a simple string, but they may be an object with a "name" and "parameter_type_overrides". Therefore, we need to use the same `Method.from_json` logic to determine the canonical method name.

Reviewed By: anwesht

Differential Revision: D89811839

fbshipit-source-id: 6c7fd0c004d51349b96848ffe790fd85062c188e