More stable output in mariana trench python parser
Summary:
While working on the the rust port, I noticed that the ordering of items in `extra_traces` was unstable and therefore hard to replicate across languages.

Let's use the third-party `OrderedSet` library to ensure the mariana trench parser outputs are stable for a given input.

We could also have sorted the outputs, but using an ordered set with predictable insertion order should be sufficient since we don't need sorted output and defining a sort order is also tricky.

Reviewed By: fahndrich

Differential Revision: D89811840

fbshipit-source-id: 36dcd20ce3dfbd26ebd3070aeffc3f639ca6bec4